% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{accumInReach}
\alias{accumInReach}
\title{Accumulates a variable within a reach}
\usage{
accumInReach(ws, x, variable = "length")
}
\arguments{
\item{ws}{A watershed}

\item{x}{A focal site}

\item{length}{The variable to accumulate}
}
\value{
A vector of accumulated values (or NA for unconnected pixels). 
Accumulated values
will be positive going downstream, negative going upstream
}
\description{
Accumulates a variable within a reach
}
