% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r, R/topology.r
\name{accumulate}
\alias{accumulate}
\title{Watershed analysis tools}
\usage{
accumulate(ws, upstream, downstream = Inf, direction = c("down", "up"),
  variable = "length", parallel = FALSE)

accumulate(ws, upstream, downstream = Inf, direction = c("down", "up"),
  variable = "length", parallel = FALSE)
}
\arguments{
\item{ws}{Watershed object}

\item{upstream}{ID number of upstream point; if Inf ALL upstream pixels will be returned}

\item{downstream}{ID of downstream point; if Inf ALL downstream pixels will be returned}

\item{direction}{In which direction should the variable be accumulated; see 'details'}

\item{variable}{The variable to accumulate}

\item{parallel}{Boolean, should parallel processing be used?}

\item{dem}{Raster or character; a digital elevation model, preferably one filled using [fillDEM()]. If specified as a character, a layer with that name from the existing [GrassSession()] given by gs will be used.}

\item{threshold}{Minimum size of an exterior watershed, in *cells*}

\item{gs}{An optional [GrassSession()]; if missing a new one will be created}

\item{outputName}{Optional character vector; if missing, a raster layer is returned, otherwise the raster is written to the grass session with this name and the modified GrassSession is returned.}

\item{file}{The file name of the raster to be returned (if no `outputName` is provided), see `details`.}

\item{...}{Additional parameters to pass to [GrassSession()]}
}
\value{
A [raster::stack()] with two layers, flow accumulation ('accumulation') and drainage
		direction ('drainage') (if outputName is missing), or a GrassSession otherwise.
		Drainage direction values can be negative (indicating water flowing out of the map region),
		zero (indicating a portion of a natural depression), or positive. Positive values are 
		measured counterclockwise from northeast; 1 is northeast, 2 north,, etc through 8 flowing #'	  due east.

A matrix, first column connected pixelIDs, second the accumulated variable
}
\description{
If direction is down, then the accumulated distances will reflect the distance from the 
`upstream` point to each downstream point, and values will be positive. If direction is "up",
then the distances will be negative and accumulate from `downstream` to `upstream`.
}
\details{
#' @details This is a wrapper for [r.watershed](https://grass.osgeo.org/grass74/manuals/r.watershed.html)

It is recommended to specify the `file` parameter (including the extension to specify
file format; e.g., .tif, .grd). If not specified, a temp file will be created and will be
lost at the end of the R session
}
