% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{accumulateOne}
\alias{accumulateOne}
\title{Accumulate a variable for one pixel to all up- and downstream pixels}
\usage{
accumulateOne(ws, x, variable = "length")
}
\arguments{
\item{ws}{A watershed}

\item{x}{A focal site}

\item{length}{The variable to accumulate}
}
\value{
a vector of accumulated values (or NA for unconnected pixels). #' Accumulated values
will be positive going downstream, negative going upstream
}
\description{
Accumulate a variable for one pixel to all up- and downstream pixels
}
