% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_methods.r
\name{as.sf.Watershed}
\alias{as.sf.Watershed}
\title{Convert a watershed to an sf LINESTRING collection}
\usage{
as.sf.Watershed(x)
}
\arguments{
\item{x}{A Watershed}
}
\value{
An sf LINESTRING
}
\description{
Currently the only attribute that is preserved is the reachID
}
