% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.r
\name{catchment}
\alias{catchment}
\title{Catchment deliniation}
\usage{
catchment(x, drainage, gs, areas, file = NULL, output = c("area",
  "raster", "sf"), ...)
}
\arguments{
\item{x}{An object of class [sp::SpatialPoints], [sf::sf], or a matrix of x-y coordinates}

\item{drainage}{Raster or character, Drainage direction raster, from e.g., [accumulate()]. If specified as a character, a layer with that name from the existing [GrassSession()] given by gs will be used.}

\item{gs}{An optional [GrassSession()]; if missing a new one will be created}

\item{areas}{DEPRECATED, use `output` instead. Logical; if `TRUE`, returns catchment area for each point, 
otherwise returns
a [raster::RasterStack], one layer per point, with each layer delimiting the catchment for its
respective point.}

\item{file}{The file name of the raster to be returned if `output='raster'`, see `details`.}

\item{output}{One of 'area', 'raster', 'sf' determining the output format}

\item{...}{Additional parameters to pass to [GrassSession()]}
}
\value{
Depends on the value of the argument `output`:
 * 'area': a vector of catchment areas, one per point in x
 * 'raster': a [rasterLayer][raster::raster()] (either a single raster or a stack, if length(x) > 1)
 * 'sf': a polygon layer of class 'sf'
}
\description{
Catchment deliniation
}
\details{
This is a wrapper for [r.water.outlet](https://grass.osgeo.org/grass74/manuals/r.water.outlet.html)

It is recommended to specify the `file` parameter (including the extension to specify
file format; e.g., .tif, .grd). If not specified, a temp file will be created and will be
lost at the end of the R session

If no `grass_session` is specified, then a new session will be created, in which case `drainage`
  is required and will be copied to the session.

This tool will perform much better if it is run with an existing grass session with 
drainage direction already added:
		`gs <- GrassSession(layer = drainage, layer_name = "drainage")`
		`cAreas <- catchment(points, drain_name = "drainage", grass_session = gs, areas = TRUE)`
	However note that one raster file is created *per point* in x, thus it is a good idea
	with many points to specify area = TRUE to get area (rather than the raster) for each point
}
