% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.r
\name{catchmentArea}
\alias{catchmentArea}
\title{Compute the area of a delineated catchment}
\usage{
catchmentArea(layer, gs)
}
\arguments{
\item{layer}{Character; the name of the grass catchment from which to compute area}

\item{gs}{A [GrassSession()]}
}
\value{
Numeric; area of the catchment
}
\description{
Compute the area of a delineated catchment
}
\keyword{internal}
