% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{cfs_cms}
\alias{cfs_cms}
\title{cfs to cms or back}
\usage{
cfs_cms(x, cfs = TRUE)
}
\arguments{
\item{x}{a continuous numeric column}

\item{cfs}{convert to cfs as default, FALSE for cms}
}
\description{
cfs to cms or back
}
