% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{confluences}
\alias{confluences}
\title{Get data from all confluences of a watershed}
\usage{
confluences(ws)
}
\arguments{
\item{ws}{Watershed object}
}
\value{
A `data.frame` containing data for all confluences
}
\description{
Get data from all confluences of a watershed
}
