% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{connect}
\alias{connect}
\title{Returns all points in the watershed connecting two points}
\usage{
connect(ws, point1, point2)
}
\arguments{
\item{ws}{Watershed object}

\item{point1}{upstream point, either an ID number of a point in the watershed, 
or an object inheriting from SpatialPoints}

\item{point2}{downstream point, either an ID number of a point in the watershed, 
or an object inheriting from SpatialPoints}
}
\description{
Returns all points in the watershed connecting two points
}
