% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{CropPtBuff}
\alias{CropPtBuff}
\title{Crop raster around a point}
\usage{
CropPtBuff(pt, ras, buff)
}
\arguments{
\item{pt}{a spatial point}

\item{ras}{a raster}

\item{buff}{distance to crop around point, in map units}
}
\value{
Cropped raster
}
\description{
Crop raster around a point
}
\keyword{internal}
