% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.r
\name{cropToCatchment}
\alias{cropToCatchment}
\title{Crops a delineated stream to the catchment at a specified point}
\usage{
cropToCatchment(x, streamRaster, streamVector, drainage, gs, file,
  trim = TRUE, ...)
}
\arguments{
\item{x}{The point at which to compute the catchment}

\item{streamRaster}{A RasterLayer stream delineation (required), or character giving the 
name of a Grass object}

\item{streamVector}{A SpatialLines or sf stream delineation (optional), or character giving the 
name of a Grass object}

\item{drainage}{Raster or character, Drainage direction raster, from e.g., [accumulate()]. 
If specified as a character, a layer with that name from the existing [GrassSession()]
given by gs will be used.}

\item{gs}{An optional [GrassSession()]; if missing a new one will be created}

\item{file}{The file name of the raster to be returned (if `areas` is `FALSE`), see `details`.}

\item{trim}{Should the resulting layers be trimmed (cropped to the non-NA extent)?}

\item{...}{Additional parameters to pass to [GrassSession()]}
}
\value{
As [extractStream()]
}
\description{
Crops a delineated stream to the catchment at a specified point
}
