% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{CV}
\alias{CV}
\title{Utility Functions}
\usage{
CV(x, na.rm = TRUE)
}
\arguments{
\item{x}{a continuous numeric column}

\item{package}{an unquoted package name}
}
\description{
A suite of functions that may or may not be useful.
}
\details{
Details: As above
}
