% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{dCdt_transport}
\alias{dCdt_transport}
\alias{dCdt_transport_r}
\title{Compute flux due to transport}
\usage{
dCdt_transport(t, y, parms)

dCdt_transport_r(t, y, parms)
}
\arguments{
\item{t}{The time step}

\item{y}{The state variable}

\item{parms}{Model parameters and data}
}
\value{
Flux vector, in units of state variable per unit time, for each spatial unit
}
\description{
Compute flux due to transport
}
\details{
At present, only advective transport is considered
}
\keyword{internal}
