% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmat}
\alias{dmat}
\title{Construct a distance matrix between one set of points and another}
\usage{
dmat(x, dsPixel, nx, value)
}
\arguments{
\item{x}{Set of pixels to which to compute distance}

\item{nx}{The total number of pixels in the network}

\item{value}{The value to be added (e.g., length) when computing distance}

\item{diPixel}{A vector of downstream pixels, diPixel[i] is downstream from pixel i}
}
\value{
A matrix with dimensions `length(x)` by `nrow(ws)`
}
\description{
Construct a distance matrix between one set of points and another
}
