% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{doStrahler}
\alias{doStrahler}
\title{Compute strahler order for pixel x}
\usage{
doStrahler(ws, x, streamOrder)
}
\arguments{
\item{ws}{A watershed}

\item{x}{A pixel id}

\item{streamOrder}{Vector of streamOrders already computed (should be NA if not done yet)}
}
\value{
A matrix, giving pixel IDs in the first column and computed stream order in the second
}
\description{
Compute strahler order for pixel x
}
\keyword{internal}
