% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{.create_reach_adjacency}
\alias{.create_reach_adjacency}
\title{Produce a reach by reach adjacency matrix}
\usage{
.create_reach_adjacency(ws)
}
\description{
Produce a reach by reach adjacency matrix
}
\keyword{internal}
