% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{.create_reach_connectivity}
\alias{.create_reach_connectivity}
\title{Produce a reach by reach connectivity matrix}
\usage{
.create_reach_connectivity(ws, self = TRUE)
}
\arguments{
\item{ws}{A watershed}

\item{self}{If TRUE, a reach is considered connected to itself
Nonzero values indicate that a reachID in a row is downstream from that column}
}
\value{
A [Matrix::sparseMatrix()]
}
\description{
Produce a reach by reach connectivity matrix
}
\keyword{internal}
