% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{.do_ws_intersect}
\alias{.do_ws_intersect}
\title{Helper function to perform the watershed intersection on a single reach}
\usage{
.do_ws_intersect(x, y, ws, width, gs, dname, riv_buff)
}
\arguments{
\item{x}{A reachID to operate on}

\item{y}{A raster stack to intersect}

\item{ws}{A watershed}

\item{width}{Buffer width}

\item{gs}{A grass session}

\item{dname}{Drainage layer in the grass session}

\item{riv_buff}{A buffer layer for the entire river}
}
\description{
Helper function to perform the watershed intersection on a single reach
}
