% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{.do_ws_zonal}
\alias{.do_ws_zonal}
\title{Compute zonal stats for a single reach}
\usage{
.do_ws_zonal(rid, ws, width, drainage, gs, riv_buff, stats)
}
\arguments{
\item{rid}{A reachID}

\item{ws}{A Watershed}

\item{width}{The width of the riparian buffer}

\item{drainage}{Drainage direction raster, or name of raster from gs}

\item{gs}{A GrassSession()}

\item{riv_buff}{Buffer for the entire river}

\item{stats}{A list of statistics function names}

\item{ras}{The raster to compute zonal stats on}
}
\description{
Compute zonal stats for a single reach
}
