% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{.make_ws_stack}
\alias{.make_ws_stack}
\title{Convert a Watershed back to a watershed-style stack}
\usage{
.make_ws_stack(ws)
}
\arguments{
\item{ws}{A watershed}
}
\value{
A raster stack following the format from the watershed package
}
\description{
Convert a Watershed back to a watershed-style stack
}
\keyword{internal}
