% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{.process_area}
\alias{.process_area}
\title{Take whatever format of areal layer, process into a raster for ws_intersect}
\usage{
.process_area(x, id, ras)
}
\arguments{
\item{x}{The areal layer to process}

\item{id}{The names of one or more columns in `x` that should be summarised along}

\item{ras}{A raster layer template to create rasters from polygons}
}
\description{
Take whatever format of areal layer, process into a raster for ws_intersect
}
