% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_methods.r
\name{.reach_to_sf}
\alias{.reach_to_sf}
\title{Produce an sf LINESTRING from a single reach}
\usage{
.reach_to_sf(x, adj)
}
\arguments{
\item{x}{data portion of a watershed, containing only a single reach}

\item{adj}{Adjacency matrix, just for the reach}
}
\description{
Produce an sf LINESTRING from a single reach
}
