% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{.reach_zone}
\alias{.reach_zone}
\title{Compute a zones for a single reach}
\usage{
.reach_zone(ws, rid, drainage, gs, width, riv_buff)
}
\arguments{
\item{ws}{A Watershed}

\item{rid}{A reachID}

\item{drainage}{Drainage direction raster, or name of raster from gs}

\item{gs}{A GrassSession()}

\item{width}{The width of the riparian buffer}

\item{riv_buff}{Buffer for the entire river}
}
\value{
list of 3 rasters; the buffer for the reach, the catchment for the reach
intersected with the buffer for the whole river, and the catchment for the reach
}
\description{
Compute a zones for a single reach
}
\keyword{internal}
