% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{.rebuild_reach_topology}
\alias{.rebuild_reach_topology}
\title{Reconstruct the reach topology}
\usage{
.rebuild_reach_topology(ws)
}
\arguments{
\item{ws}{A watershed}
}
\value{
A modified watershed
}
\description{
After editing a watershed's pixels or reaches, the reach topology will need to be rebuilt using this function.
}
\keyword{internal}
