% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{.rebuild_reaches}
\alias{.rebuild_reaches}
\title{Prepare watershed by setting up proper reach structures}
\usage{
.rebuild_reaches(x)
}
\arguments{
\item{x}{A watershed}
}
\value{
A modified watershed with reaches properly set
}
\description{
Prepare watershed by setting up proper reach structures
}
\keyword{internal}
