% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{.renumber_reaches}
\alias{.renumber_reaches}
\title{Renumber reaches, starting with one at the headwaters}
\usage{
.renumber_reaches(ws)
}
\arguments{
\item{ws}{A watershed}
}
\description{
Renumber reaches, starting with one at the headwaters
}
