% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{.resize_reach}
\alias{.resize_reach}
\title{Resize a single reach}
\usage{
.resize_reach(adj, len, size, min_size, start = 1)
}
\arguments{
\item{adj}{subsetted adjacency matrix from a watershed, with just the entries for a single reach}

\item{len}{length of each pixel}

\item{size}{Reach size; the desired reach size, see 'details'}

\item{min_size}{Minimum reach size}

\item{start}{The reach number to start with}
}
\description{
Resize a single reach
}
\details{
reach numbers will start from 1
}
