% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{.riv_buff_intersect}
\alias{.riv_buff_intersect}
\title{Helper function for creating river buffers}
\usage{
.riv_buff_intersect(x, y, rip_buffer, gs)
}
\arguments{
\item{x}{A river network; this should be a vector gis feature, either from the `sf` or `sp` packages}

\item{y}{The polygon feature to intersect, sf class}

\item{rip_buffer}{The width of the riparian buffer}

\item{gs}{A grass session, if missing or NA, then sf functions will be used}
}
\value{
If gs is null or grass_output is "sf", then a buffer polygon layer of class "sf", if grass_output is "string",
then the name of the vector layer in the grass session
}
\description{
Helper function for creating river buffers
}
\keyword{internal}
