% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{.sf_summary}
\alias{.sf_summary}
\title{Summarise an sf polygon feature by attributes}
\usage{
.sf_summary(x, by, subunit, ...)
}
\arguments{
\item{x}{An sf polygon}

\item{by}{Character vector, column names giving factors to summarise along. Partial matching with regular expressions
is allowed, as long as the result is unique.}

\item{subunit}{Optional character vector; If included, it will be used as a grouble variable(s) when computing proportions;
variables here will also be used as grouping variables as in 'by'}

\item{...}{Additional named arguments, see 'details'}
}
\value{
A data.table with the following columns:
* ... ID columns, same names as `by`, giving the id levels of each summarised feature
}
\description{
Summarise an sf polygon feature by attributes
}
\details{
Additional arguments can be supplied to add columns to the output. For example, adding `category = 2` will add a column
named category to the output table, with a value of 2 for all entries. Useful for example when calling .sf_summary from mapply,
and it is desirable to know the combinations in mapply resulting in a given line of output
}
\keyword{internal}
