% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{.ws_buffer}
\alias{.ws_buffer}
\title{Compute a riparian buffer around a Watershed or a reach}
\usage{
.ws_buffer(ws, buff, rid)
}
\arguments{
\item{ws}{A Watershed}

\item{buff}{The buffer width}

\item{rid}{A reach id, if not missing, the buffer will only include this reach}
}
\value{
A raster
}
\description{
Compute a riparian buffer around a Watershed or a reach
}
\keyword{internal}
