% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{downstreamDist}
\alias{downstreamDist}
\title{Build a downstream distance matrix for a list of sites}
\usage{
downstreamDist(ws, x, variable = "length", fun = sum)
}
\arguments{
\item{ws}{A watershed}

\item{x}{A vector of pixel ids}

\item{variable}{The variable to use for the distance metric}

\item{fun}{The function to apply between each pair of sites}
}
\value{
A site by site distance [Matrix::sparseMatrix()]
}
\description{
For each site, this function identifies the other sites that are downstream of it and computes
a 'distance' between each site and all of its downstream sites. By default, this is river,
distance, computed by summing the length of stream between the two sites. For any entry in
this distance matrix [i,j], the value is either a nonzero number giving the downstream
distance from site i to site j, or zero, indicating that j is not downstream of i.
}
\details{
For a more general (and much slower) distance computation, see [wsDistance()]
}
