% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{downstreamPixelIds}
\alias{downstreamPixelIds}
\title{Get the pixel ID of the next downstream pixel for each pixel in the watershed}
\usage{
downstreamPixelIds(ws)
}
\arguments{
\item{ws}{A watershed object}
}
\value{
A vector of pixel IDs
}
\description{
Get the pixel ID of the next downstream pixel for each pixel in the watershed
}
