% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dowy.R
\name{dowy}
\alias{dowy}
\title{Add Day of Water Year}
\usage{
dowy(YYYYMMDD_HMS)
}
\arguments{
\item{YYYYMMDD_HMS}{a POSIXct formatted column}
}
\description{
This function takes a \emph{date} or \emph{datetime} column and
returns the Day of Water Year (from Oct 1 to Sep 30).
}
\details{
Details: As above
}
