% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{extract}
\alias{extract}
\title{Extract pixelIDs from a watershed at spatial locations}
\usage{
extract(ws, x)
}
\arguments{
\item{ws}{A watershed object}

\item{x}{An object inheriting from `sp::SpatialPoints()`}
}
\value{
A vector of pixel IDs
}
\description{
Extract pixelIDs from a watershed at spatial locations
}
