% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{extractStream}
\alias{extractStream}
\title{Produce a map of a stream network}
\usage{
extractStream(
  dem,
  accumulation,
  threshold,
  qthresh = 0.95,
  weights,
  gs,
  outputName,
  type = c("raster", "vector", "both"),
  file = NULL,
  ...
)
}
\arguments{
\item{dem}{Raster or character; a digital elevation model, preferably one filled using [fillDEM()]. If specified as a character, a layer with that name from the existing [GrassSession()] given by gs will be used.}

\item{accumulation}{Raster or character; flow accumulation later; see details.}

\item{threshold}{Accumulation threshold; optional, see details}

\item{qthresh}{Quantile threshold to use if `threshold` is missing}

\item{weights}{A raster layer specifying the weights to apply to flow accumulation before comparing to `threshold`.}

\item{gs}{An optional [GrassSession()]; if missing a new one will be created}

\item{outputName}{Optional character; if missing, a raster layer is returned, otherwise the raster is written to the grass session with this name and the modified GrassSession is returned.}

\item{file}{The file name of the raster to be returned (if no `outputName` is provided), see `details`.}

\item{...}{Additional parameters to pass to [GrassSession()] or r.stream.extract}
}
\value{
If `outputName` is missing and `type=='raster'`; a [raster::raster]; if 
  `outputName` is missing and `type=='vector'`; a [sp::SpatialLinesDataFrame];
  otherwuse a `GrassSession`
}
\description{
Produce a map of a stream network
}
\details{
A flow accumulation raster can be provided (as a rasterLayer or as the name of a raster in an existing Grass Session). If not provided, it will be computed.

If no threshold is used, then one will be set automatically using the quantiles of the accumulation raster

This is a wrapper for [r.stream.extract](https://grass.osgeo.org/grass74/manuals/r.stream.extract.html)

It is recommended to specify the `file` parameter (including the extension to specify
file format; e.g., .tif, .grd). If not specified, a temp file will be created and will be
lost at the end of the R session
}
