% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{fillDEM}
\alias{fillDEM}
\title{Produce a DEM with basins filled in, as well as a flow direction raster}
\usage{
fillDEM(dem, gs, filledDEM, probs, file = NULL, ...)
}
\arguments{
\item{dem}{Raster or character; a digital elevation model. If specified as a character, a layer with that name from the existing [GrassSession()] given by gs will be used.}

\item{gs}{An optional [GrassSession()]; if missing a new one will be created}

\item{filledDEM}{Optional character; if missing, a raster stack is returned, otherwise the filled DEM is written to the grass session with this name and the modified GrassSession is returned.}

\item{probs}{Optional character; as with filled DEM, for problem areas}

\item{file}{The file name of the raster to be returned (if no `outputName` is provided), see `details`.}

\item{...}{Additional parameters to pass to [GrassSession()]}
}
\value{
A [raster::stack], optionally written to `file` with 2 layers: the filled DEM and a list of problem areas (if outputName is missing), or a GrassSession otherwise
}
\description{
Produce a DEM with basins filled in, as well as a flow direction raster
}
\details{
This is a wrapper for [r.fill.dir](https://grass.osgeo.org/grass74/manuals/r.fill.dir.html)

It is recommended to specify the `file` parameter (including the extension to specify
file format; e.g., .tif, .grd). If not specified, a temp file will be created and will be
lost at the end of the R session
}
