% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{findClosest}
\alias{findClosest}
\title{Find the closest non NA point in a raster}
\usage{
findClosest(x, y, buff)
}
\arguments{
\item{x}{a spatial object}

\item{y}{a raster}

\item{buff}{The distance (in map units) to restrict the search from points in x}
}
\value{
The index in y of the closest point
}
\description{
Find the closest non NA point in a raster
}
\details{
Finds the points in y that are closest to x
}
\keyword{internal}
