% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basins.R
\name{get_basins}
\alias{get_basins}
\title{Get HUC Basins}
\usage{
get_basins(sites, filePath = NA)
}
\arguments{
\item{sites}{an sf formatted object}

\item{filePath}{a location to save the file, if specified}
}
\description{
This function uses a spatial polygon/sf object to search and download the watersheds for a given area.
Based on USGS tools and using NHD Data.
}
\details{
This function requires an internet connection, as well as the \code{httr} and \code{sf} packages.
}
\examples{

# get data
yose <- data.frame(site=c("YOSEMITE","LOON") ,lon = c(-119.76865,-120.31862), lat = c(37.95748, 38.99134))
yose_sf <- sf::st_as_sf(yose, coords=c("lon", "lat"), crs=4326, remove=FALSE)
# now run function:


}
