% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCDEC.R
\name{get_cdec}
\alias{get_cdec}
\title{Get CDEC Data}
\usage{
get_cdec(station, sensor, duration, start, end)
}
\arguments{
\item{station}{# Station is 3 letter abbreviation (see \url{https://info.water.ca.gov/staMeta.html})}

\item{sensor}{# Sensor is number.}

\item{duration}{# Duration is E=event, D=Daily, H=Hourly.}

\item{start}{# A start date formatted as "YYYY-MM-DD".}

\item{end}{# A end date formatted as "YYYY-MM-DD".}
}
\value{
Will return a dataframe with station ID and associated sensor data.
}
\description{
This function takes a \emph{station} and \emph{sensor} along
with a \emph{duration} and \emph{start} and \emph{end} date.
It returns associated CDEC sensor data in a dataframe.
This function can download data for a single station or
a list of multiple stations at once using \strong{purrr}.
}
\details{
Commonly used sensors include:
 \itemize{
   \item 1  stage (ft)
   \item 20 flow (cfs)
   \item 2  rain accum (in)
   \item 16 precip tippingbucket (in)
   \item 45 ppt incremental (in)
   \item 3  snow water content (in)
   \item 18 snow depth (in)
   \item 6  reservoir elevation (ft)
   \item 15 reservoir storage (ac-ft)
   \item 76 reservoir inflow
   \item 25 water temp
   \item 4  air temp
   }
}
\examples{
# set up the parameters to use in function:
sens <- c(20)
startT <- c("2018-12-01")
endT <- c("2019-03-01")
stations <- c("CLW", "MLW", "TIS")
dur  <- c("H") # hourly here, try this with "H" "D" or "E"
# combine into one argument list (note, order matches function order)
varList <- list(stations, sens, dur, startT, endT)
# now run function:
# use PURRR package to map each argument in parallel, then combine
cdec_hrly <- pmap(varList, get_cdec) \%>\% bind_rows()
# check records by station
table(cdec_hrly$station_id)
# check date range
cdec_hrly \%>\%
  group_by(station_id) \%>\%
  summarize(mx=max(datetime), mn=min(datetime))
# Format to Daily Data (requires dplyr)
cdec <- cdec_hrly \%>\%
  mutate(date = as.Date(datetime)) \%>\%
  group_by(station_id, date) \%>\%
  summarize("daily_cfs" = mean(value, na.rm = T)) \%>\%
  filter(!is.na(daily_cfs))
}
\seealso{
\itemize{
  \item To see a list of Real-Time Stations: \url{http://cdec.water.ca.gov/misc/realStations.html}
  \item To see a list of Daily Stations: \url{http://cdec.water.ca.gov/misc/dailyStations.html}
  \item To see a list of sensors:  \url{http://cdec.water.ca.gov/misc/senslist.html}
  }
}
