% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shp_zip.R
\name{get_shp_zip}
\alias{get_shp_zip}
\title{Read a zipped shapefile}
\usage{
get_shp_zip(shp_url)
}
\arguments{
\item{shp_url}{A character path to a zipped shapefile, locally or online.}
}
\value{
Returns an sf data.frame.
}
\description{
This function takes a \emph{path} to a given shapefile
from online or local directory, and
returns a \emph{sf} based data.frame to the R environment.
}
\examples{
# set up the parameters to use in function:
library(sf)
library(utils)
library(mapview)

# get a shapefile of all USGS gages in North America
shp_url <- "https://water.usgs.gov/GIS/dsdl/gagesII_9322_point_shapefile.zip"

gages2 <- get_shp_zip(shp_url)

# now plot
mapview::mapview(gages2)
}
