% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getupstreamflowlines}
\alias{getupstreamflowlines}
\title{getupstreamflowlines}
\usage{
getupstreamflowlines(lon, lat)
}
\arguments{
\item{lon}{start point longitude}

\item{lat}{start point longitude}
}
\value{
Retrieves upstream NHDplus flowlines
}
\description{
getupstreamflowlines
}
