% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{GrassSession}
\alias{GrassSession}
\title{Set up a GRASS session}
\usage{
GrassSession(layer, gisBase, layerName = NA, home = tempdir(),
  gisDbase = home, location = "NSmetabolism", mapset = "PERMANENT",
  override)
}
\arguments{
\item{layer}{A [raster::raster] object (recommended); or any other object which has `extent()` and `proj4string()` methods defined.}

\item{gisBase}{character; the location of the GRASS installation (see `details`)}

\item{layerName}{NA or character; if NA (the default), the layer will not be added to the grass session, otherwise it will be appended with this name.}

\item{home}{Location to write GRASS settings, `details`.}

\item{gisDbase}{Location to write GRASS GIS datasets; see `details`.}

\item{location}{Grass location name}

\item{mapset}{Grass mapset name}

\item{override}{Logical;  see `details`}
}
\value{
An S3 [GrassSession] object
}
\description{
Set up a GRASS session
}
\details{
if `gisBase` is not provided, it can be automatically deduced in some cases.
On some systems, you can run `grass74 --config path` from the command line to get this path.

The extent, projection, and resolution of the grass session will be determined by `layer`.

by default `override` will be TRUE if home and gisDbase are set to their defaults, otherwise FALSE. If TRUE, the new session will override any existing grass session (possibly damaging/overwriting existing files). It is an error if override is FALSE and there is an already running session.
}
