% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{GSAddRaster}
\alias{GSAddRaster}
\title{Add a RasterLayer to a grass session and return the modified session}
\usage{
GSAddRaster(x, layerName, gs, overwrite = TRUE)
}
\arguments{
\item{x}{A [raster::raster] object or a [sp::SpatialGridDataFrame]}

\item{layerName}{character; the name of the layer to add to grass.}

\item{gs}{A [GrassSession] object}
}
\value{
An S3 [GrassSession] object
}
\description{
Add a RasterLayer to a grass session and return the modified session
}
