% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{GSAppendRasterName}
\alias{GSAppendRasterName}
\title{Add a RasterLayer *name* to a grass session and return the modified session}
\usage{
GSAppendRasterName(x, gs)
}
\arguments{
\item{x}{Character; the name of the layer(s) to add}

\item{gs}{A [GrassSession] object}
}
\value{
An S3 [GrassSession] object
}
\description{
Add a RasterLayer *name* to a grass session and return the modified session
}
\keyword{internal}
