% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{GSClean}
\alias{GSClean}
\title{Delete layers from a GrassSession}
\usage{
GSClean(layer, gs, type = c("raster", "vector"))
}
\arguments{
\item{layer}{Layer name to delete}

\item{gs}{GrassSession to operate on}

\item{type}{Type of layer, either raster of vector}
}
\value{
Modified grass session
}
\description{
Delete layers from a GrassSession
}
\keyword{internal}
