% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{GSGetRaster}
\alias{GSGetRaster}
\title{Read a raster from a grass session}
\usage{
GSGetRaster(layer, gs, file)
}
\arguments{
\item{layer}{The name of the layer(s) to read}

\item{gs}{A [GrassSession] object}

\item{file}{(optional, recommended) Where to store the raster on disk}
}
\value{
A raster or raster stack
}
\description{
Read a raster from a grass session
}
\details{
If no file is specified, the raster will be stored in memory or as a temporary
  file and will be lost at the end of the R sessionl.
}
