% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{GSRastToPoly}
\alias{GSRastToPoly}
\title{Convert a raster to a polygon in grass}
\usage{
GSRastToPoly(rast, vect, gs)
}
\arguments{
\item{rast}{Either a rasterlayer or a file name of a raster in grass}

\item{vect}{Optional layer name to save the polygon}

\item{gs}{GrassSession to operate on}
}
\value{
If vect is missing, a spatialPolygons object, otherwise NULL (with the side-effect)
		of writing a polygon to the grass session
}
\description{
Convert a raster to a polygon in grass
}
\keyword{internal}
