% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{headwaters}
\alias{headwaters}
\title{Get data from all headwaters of a watershed}
\usage{
headwaters(ws)
}
\arguments{
\item{ws}{Watershed object}
}
\value{
a `data.frame` containing data for all headwaters
}
\description{
Get data from all headwaters of a watershed
}
