% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{HUC12_Outlet_COMIDs_CONUS}
\alias{HUC12_Outlet_COMIDs_CONUS}
\title{HUC12 and NHDPlus Version 2 COMID Associations}
\format{
An object of class \code{data.frame} with 81947 rows and 6 columns.
}
\usage{
HUC12_Outlet_COMIDs_CONUS
}
\value{
The dataset HUC12_COMID_PU_CONUS.dbf contains HUC12 associations
for every NHDPlus Version 2 reach except for Coastlines.
}
\description{
HUC12 and NHDPlus Version 2 COMID Associations
}
\keyword{datasets}
