% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{hydraulic_geometry}
\alias{hydraulic_geometry}
\title{Compute hydraulic geometry from discharge}
\usage{
hydraulic_geometry(discharge, pars)
}
\arguments{
\item{discharge}{A vector of discharge values.}

\item{pars}{Optional parameters for the scaling relationships; see 'details'}
}
\value{
A data frame with discharge, velocity, depth, and width
}
\description{
Compute hydraulic geometry from discharge
}
\details{
If given, `pars` should be a named `list` with three elements: `velocity`, `depth`
		and `width`. Each element must be a vector with two elements, the first is the intercept
		and the second the slope. If omitted, values from Raymond et al will be used.

For the Raymond et al parameters, discharge should be in units of \eqn{m^3 s^{-1}}. 
		All other variables are in meters.
}
\references{
Raymond, PA et al. 2012. Scaling the gas transfer velocity and hydraulic 
		geometry in streams and small rivers. *Limnology and Oceanography: Fluids and
		Environments*. **2**:41-53.
}
