% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.r
\name{interpolate}
\alias{interpolate}
\title{Provide interpolation in time}
\usage{
interpolate(ws, values, times_out, sites = as.integer(rownames(values)),
  times = as.integer(colnames(values)), ...)
}
\arguments{
\item{ws}{A watershed}

\item{values}{A matrix, with sites in rows and times in columns.}

\item{times_out}{A vector of target times for interpolation}

\item{sites}{An optional vector of sites with length = nrow(values).}

\item{times}{An optional vector of times at which observations have been made.}

\item{...}{Additional parameters to pass to approx}
}
\description{
Provide interpolation in time
}
