% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadpackages}
\alias{loadpackages}
\title{loadpackages}
\usage{
loadpackages(packages)
}
\arguments{
\item{packages}{\link{character} character vector of packages to call with library()}
}
\value{
Loads multiple packages
}
\description{
loadpackages
}
