% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_methods.r
\name{[.Watershed}
\alias{[.Watershed}
\alias{names.Watershed}
\alias{head.Watershed}
\alias{summary.Watershed}
\alias{plot.Watershed}
\title{S3 methods for watersheds}
\usage{
\method{[}{Watershed}(x, i, j, ...)

\method{names}{Watershed}(x)

head.Watershed(x, ...)

\method{summary}{Watershed}(x, ...)

\method{plot}{Watershed}(x, variable, transform, size = 0.5, xlim, ylim, viridis_col = "inferno")
}
\arguments{
\item{x}{A watershed object}

\item{i}{A vector of pixel IDs}

\item{j}{Variable names or column numbers to extract}

\item{...}{Additional parameters to pass to other methods 
(usually from [data.frame] or [SpatialPointsDataFrame])}

\item{variable}{Optional, name of variable to plot}

\item{transform}{A transformation function (e.g., `log`) to apply to the variable}

\item{size}{Size of the plotted points}

\item{xlim}{Optional x-axis limits}

\item{ylim}{Optional y-axis limits}
}
\value{
A data.frame of the watershed data

A vector of variable names

As [head.data.frame()]

As [summary.data.frame()]
}
\description{
S3 methods for watersheds
}
