% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiply_sparse}
\alias{multiply_sparse}
\title{Quick and dirty vector by sparse matrix multiplication
note that no bounds checking is done}
\arguments{
\item{mat}{A 3-column matrix giving the rows and columns of the non-zero entries of the 
sparse matrix (columns 1 and 2) and the value (column 3)}

\item{y}{The vector by which to multiply mat}
}
\description{
Quick and dirty vector by sparse matrix multiplication
note that no bounds checking is done
}
\keyword{internal}
