% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{nearestDownstreamNeighbor}
\alias{nearestDownstreamNeighbor}
\title{Construct a matrix identifying the nearest downstream neighbor from a list of sites}
\usage{
nearestDownstreamNeighbor(ws, x, siteNames)
}
\arguments{
\item{ws}{A watershed}

\item{x}{A vector of pixel IDs}

\item{names}{Optional vector of site names}
}
\value{
A 2-column matrix; the first column gives the ID of a pixel, the second its nearest
downstream neighbor. Pixels in `x` that have no nearest neighbor are excluded.
If 'names' is specified, the output matrix will give site names instead of pixelIDs
}
\description{
Construct a matrix identifying the nearest downstream neighbor from a list of sites
}
