% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{nearestNeighbors}
\alias{nearestNeighbors}
\title{Find the nearest neighbors for a focal pixel among a list of sites}
\usage{
nearestNeighbors(ws, x, distMatrix, sites, selfAdjacency = FALSE)
}
\arguments{
\item{ws}{A watershed, optional unless `distMatrix` is missing}

\item{x}{A vector of focal pixels (always required)}

\item{distMatrix}{A site by pixel distance matrix, will be computed if not specified}

\item{sites}{A list of sites for the distance matrix, required if distMatrix is missing}

\item{selfAdjacency}{if TRUE, a site can be downstream of itself}
}
\value{
A list with a downstream and an upstream element. Each element is a matrix, where
the first column corresponds to the focal pixels in `x`, and the second column lists the
nearest neighbors in `sites`.
}
\description{
For a list of `sites`, this function first computes a distance matrix(if not specified),
then finds the nearest upstream and downstream neighbors for focal pixels `x`. Note that
`x` does not necessarily have to be in the list of sites, they can be anywhere in the watershed.
}
