% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{outlets}
\alias{outlets}
\title{Get data from all outlets of a watershed}
\usage{
outlets(ws)
}
\arguments{
\item{ws}{Watershed object}
}
\value{
a `data.frame` containing data for all outlets
}
\description{
Get data from all outlets of a watershed
}
