% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owin.R
\name{owin}
\alias{owin}
\title{Open Directory on PC}
\usage{
owin(...)
}
\arguments{
\item{...}{empty space}
}
\description{
This is a helper function to open current project directory
}
\details{
It doesn't require any specific parameters
}
