% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{plot.Watershed}
\alias{plot.Watershed}
\title{Plot method for a [Watershed()]}
\usage{
\method{plot}{Watershed}(x, variable, transform, size = 0.5)
}
\arguments{
\item{x}{Watershed object to plot}

\item{variable}{Optional, name of variable to plot}

\item{transform}{A transformation function (e.g., `log`) to apply to the variable}

\item{size}{Size of the plotted points}
}
\description{
Plot method for a [Watershed()]
}
