% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{q_from_flow}
\alias{q_from_flow}
\title{Compute discharge from flow measurements}
\usage{
q_from_flow(depth, velocity, flowID = rep(1, length(depth)), width,
  widthID = rep(1, length(width)), na.rm = FALSE)
}
\arguments{
\item{depth}{a vector of depth measurements}

\item{velocity}{a vector of velocity measurements}

\item{flowID}{a vector of IDs (e.g., site numbers) corresponding to depth/velocity measurements}

\item{width}{a vector of stream widths}

\item{widthID}{a vector of IDs for width measurements to match to flowID}

\item{na.rm}{should NAs be removed before computation?}
}
\value{
A data.table giving discharge by ID
}
\description{
Compute discharge from flow measurements
}
\keyword{internal}
