% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{reachAdj}
\alias{reachAdj}
\title{Find all reaches directly upstream from a given reach}
\usage{
reachAdj(ws, rch)
}
\description{
Find all reaches directly upstream from a given reach
}
\keyword{internal}
