% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{renumberReaches}
\alias{renumberReaches}
\title{Renumber reachIDs to start at 1 and increase one at a time, preserving ordering}
\usage{
renumberReaches(rIDs)
}
\description{
Renumber reachIDs to start at 1 and increase one at a time, preserving ordering
}
\keyword{internal}
