% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{resize_reaches}
\alias{resize_reaches}
\title{Resize a watershed's reaches}
\usage{
resize_reaches(ws, size, min_size, parallel = TRUE)
}
\arguments{
\item{ws}{A watershed}

\item{size}{Reach size; the desired reach size, see 'details'}

\item{min_size}{Minimum reach size}

\item{parallel}{Use the parallel package to speed computation on mac/linux?}
}
\value{
A modified watershed
}
\description{
Resize a watershed's reaches
}
\details{
By default watersheds are created with simple reaches that stretch between confluences. These can be 
quite long, however, so this function will resize them, starting at headwaters.

The size parameter only gives a target size, reaches can be smaller or larger in specific circumstances. Resizing
starts at headwaters, When a confluence is reached, the final reach before the confluence will often be smaller than
`size`. Thus, it is also necessary to specify `min_size`. If the reach upstream of a confluence is smaller than 
`min_size`, the reach will be incorporated into the next reach upstream. If there is no reach upstream, the reach will
be deleted. If it is larger than `min_size`, it will be left as a separate reach. Thus, the largest reaches will be
`size + min_size` in length, and the smallest will be `min_size`
}
