% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riv8_runoff}
\alias{riv8_runoff}
\title{River Runoff from 8-station index.}
\format{
A data frame with 115 rows (updated annually) and 7 variables:
\describe{
  \item{WY}{the water year}
  \item{Dec}{The 8 station index for Dec of a given water year}
  \item{Jan}{The 8 station index for Jan of a given water year}
  \item{Feb}{The 8 station index for Feb of a given water year}
  \item{Mar}{The 8 station index for Mar of a given water year}
  \item{Apr}{The 8 station index for Apr of a given water year}
  \item{May}{The 8 station index for May of a given water year}
}
}
\source{
\url{https://cdec.water.ca.gov/reportapp/javareports?name=PLOT_ESI.pdf}
}
\usage{
riv8_runoff
}
\description{
The 8-station index for CA river runoff.
}
\keyword{datasets}
