% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{siteByPixel}
\alias{siteByPixel}
\title{Compute a site by pixel accumulation matrix}
\usage{
siteByPixel(ws, x, variable = "length")
}
\arguments{
\item{ws}{A Watershed}

\item{x}{A vector of pixel ids from which to compute the distance}

\item{variable}{The variable to use for the distance}
}
\value{
A matrix with dimensions `length(x)` by `nrow(ws)`
}
\description{
The default behavior computes distance, where positive numbers indicate downstream
distances and negative numbers indicate upstream distances. Other variables can also
be used, but in all cases the values will be summed to compute the 'distance'
}
\details{
Upstream distances do NOT include intermediate pixels; they only include pixels in `x`
}
