% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{snapToStream}
\alias{snapToStream}
\title{Snap points to a stream network raster}
\usage{
snapToStream(x, stream, buff)
}
\arguments{
\item{x}{SpatialPoints or similar object}

\item{stream}{Raster or character; a stream network raster (e.g., from [extractStream()]). 
If specified as a character, a layer with that name from the existing [GrassSession()] 
given by gs will be used.}

\item{buff}{The distance (in meters if x is in geographic coordinates, map units otherwise) to restrict the search from points in x}
}
\value{
A SpatialPointsDataFrame, with the attributes from `x` and new coordinates
}
\description{
Snap points to a stream network raster
}
\details{
If buff is too small, a nearby stream may not be found, in which case the original coordinates are returned
}
