% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{strahler}
\alias{strahler}
\title{Compute strahler stream order}
\usage{
strahler(ws, parallel = requireNamespace(parallel))
}
\arguments{
\item{ws}{A watershed}

\item{parallel}{If TRUE, will compute in parallel to speed things up. Defaults to TRUE
if parallel package is installed}
}
\value{
a vector, with one element per pixel in the watershed.
}
\description{
Compute strahler stream order
}
\details{
Computes Strahler stream order
}
