% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{stream_order}
\alias{stream_order}
\title{Compute stream order}
\usage{
stream_order(
  stream,
  drainage,
  elevation,
  gs,
  outputName,
  type = c("strahler", "horton", "shreve", "hack", "topo"),
  file,
  install = FALSE,
  force = FALSE,
  ...
)
}
\arguments{
\item{stream}{Raster or character; stream raster.}

\item{drainage}{Raster or character; drainage direction raster.}

\item{elevation}{Raster or character; a digital elevation model.}

\item{gs}{An optional [GrassSession()]; if missing a new one will be created}

\item{outputName}{The name of the output raster to create; see 'details'}

\item{type}{Type of stream order to compute; see help for r.stream.order}

\item{file}{Optional file name if a raster is to be returned}

\item{install}{Boolean, should we install the r.stream.order extension? If you get an error
that the command was not found, try setting this to TRUE the first time you run `stream_order`}

\item{force}{Force this broken function to try anyway}

\item{...}{Additional parameters to be passed to [GrassSession()]}
}
\value{
A rasterLayer or a grassSession, depending on the value of `outputName`
}
\description{
Compute stream order
}
\details{
Deprecated function, use strahler instead, this one doesn't work. If you really
		want this one, use force = TRUE

This is a wrapper for r.stream.order in GRASS GIS.

Input rasters `stream`, `drainage` and `elevation` may be either [raster::rasterLayer()]
objects or a `character`. In the former case, these layers will be added to the grass
session. In the latter, it is required that `gs` be specified and that a layer with that
name exists in the grass session.

If outputName is missing, then the function will return the stream order raster. Otherwise,
this outputName will be used to save the raster in the existing GrassSession, which will
be returned to the user
}
