% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcatchment.r
\name{subcatchment}
\alias{subcatchment}
\title{Extract a complete subcatchment from the watershed}
\usage{
subcatchment(ws, x)
}
\arguments{
\item{ws}{A [Watershed()].}

\item{x}{The point to make the new outlet for the extracted watershed.}
}
\value{
A watershed.
}
\description{
This function returns a complete subset watershed object, including all triburaties upstream
from a specified point. PixelIDs and reachIDs will be renumbered.
}
