% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{transport}
\alias{transport}
\title{Runs a transport-only model}
\usage{
transport(ws, initial, lateral, times, method = c("euler", "lsoda"),
  dt = 1)
}
\arguments{
\item{ws}{A [Watershed] object}

\item{initial}{A vector of initial conditions}

\item{lateral}{Vector of values of the state variable for lateral input for each pixel}

\item{times}{The times at which to compute the state variable}

\item{method}{The integration method to use; default is euler}

\item{dt}{The size of the time step for integration (euler method only)}
}
\value{
A state variable matrix (with one column per time step)
}
\description{
Runs a transport-only model
}
\details{
Because the appropriate time step will vary based on the units and the specific
   problem, the recommended approach is to first try a single run using lsoda (perhaps on
   a small subset), then run euler at varying time steps to find the dt that produces
   acceptable results.
}
