% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{transport_euler}
\alias{transport_euler}
\title{Do euler integration for the transport model}
\usage{
transport_euler(initial, parms, times, dt, rxn = NULL, rxnParams = list())
}
\arguments{
\item{initial}{A vector of initial conditions}

\item{parms}{Integration parameters and data}

\item{times}{The times at which to compute the state variable}

\item{dt}{The size of the time step for integration}

\item{rxn}{A function giving the time derivative of the reaction component}

\item{rxnParams}{A list of parameters to pass to `rxn`}
}
\value{
A state variable matrix (with one column per time step)
}
\description{
Do euler integration for the transport model
}
\keyword{internal}
