% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach.r
\name{trim_reaches}
\alias{trim_reaches}
\title{Remove short headwater reaches below a certain threshold}
\usage{
trim_reaches(ws, size, rebuild = TRUE)
}
\arguments{
\item{ws}{A watershed}

\item{size}{The minimum size of a reach to retain, in map units}

\item{rebuild}{Boolean; should topology be rebuilt? Normally yes, but can set to FALSE if it will be rebuilt later}
}
\value{
A modified watershed
}
\description{
Remove short headwater reaches below a certain threshold
}
