% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershedtools-package.R
\docType{package}
\name{watershedtools-package}
\alias{watershedtools}
\alias{watershedtools-package}
\title{watershedtools: Hydrologic Analysis Utilities Built on NHDPlus}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides HUC code validation and hierarchy resolution, watershed boundary retrieval, upstream and downstream stream tracing via 'NHDPlus' flow direction, pour point snapping, and feature-watershed spatial overlay. Built on 'nhdplusTools' for data access and 'sf' for spatial operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chrislyonsKY/watershedtools}
  \item Report bugs at \url{https://github.com/chrislyonsKY/watershedtools/issues}
}

}
\author{
\strong{Maintainer}: Chris Lyons \email{chris.lyons@ky.gov}

}
\keyword{internal}
