% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{ws_intersect}
\alias{ws_intersect}
\title{Intersect a river feature with a polygon feature}
\usage{
ws_intersect(ws, areas, area_id, rip_buffer = 50, drainage = NA, ...)
}
\arguments{
\item{ws}{A Watershed}

\item{areas}{A named list of areal features to summarize; should be polygon features from `sf` or `sp`}

\item{area_id}{Character vector; the names of one or more columns in `areas` that should be summarised along
the river feature; partial matching with regular expressions is allowed, but each item in `area_id` must identify 
a single column in `areas`}

\item{rip_buffer}{The width of the riparian buffer}

\item{drainage}{Optional drainage direction raster for calculating catchment; see 'details'}

\item{...}{Additional parameters to pass to [GrassSession()]}
}
\value{
A data.table summarising the polygons in each layer in areas along the river provided in x.
}
\description{
ws_intersect handles the use case where we want to summarize areal features, such as land use, soils, etc, across
reaches of a river, while taking into account drainage area, flow, etc.
}
\details{
There are three potential outputs, which will be identified in the output by the 'method' column
* `riparian`: The layers in `areas` are summarized within a riparian buffer surrounding each reach; 
 		provided if `rip_buffer` is not NA.
* `riparian_upstream`: The layers in `areas` are summarized within a riparian buffer for each focal reach and all
         upstream reaches; provided if `rip_buffer` is not NA and `drainage` is not NA
* `catchment`: The layers in `areas` are summarized within the entire catchment for each reach, provided if
         `drainage` is not NA
}
