% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{ws_slope}
\alias{ws_slope}
\title{Compute the slope along the river channel}
\usage{
ws_slope(ws, by = c("reach", "pixel"))
}
\arguments{
\item{ws}{A watershed}

\item{by}{Should the slope be computed for each reach (recommended) or from one pixel to the 
next}
}
\value{
A vector of slopes, one per pixel
}
\description{
This computes the change in elevation from one watershed unit to the next. Computation by
pixel is only recommended if the elevation model used to create the watershed is very
accurate; otherwise unrealistic slopes can occur. Reach-based computation uses a much
larger range of pixels to compute the slope, and so is less prone to DEM errors.
}
\examples{
\donttest{
ws_slope(ws)
ws_slope(ws, by = "pixel")
}
}
