% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{ws_zonal}
\alias{ws_zonal}
\title{Compute zonal statistics on a watershed}
\usage{
ws_zonal(ws, ras, statistics = list("mean", "sd"), rip_buffer = 50, ...)
}
\arguments{
\item{ws}{A Watershed]}

\item{ras}{A raster}

\item{statistics}{A list of function names; each function must accept a vector and return
a single value, and must have an na.rm argument}

\item{rip_buffer}{The width of the riparian buffer}

\item{...}{Additional parameters to pass to [GrassSession()]}
}
\value{
A data.table summarising the values in each layer in areas along the river provided in x.
}
\description{
Compute zonal statistics, where the zones are determined by watershed reaches.
}
\details{
There are three output types:
* `riparian`: The layers in `areas` are summarized within a riparian buffer surrounding each 
reach
* `riparian_upstream`: The layers in `areas` are summarized within a riparian buffer for each 
focal reach and all upstream reaches
* `catchment`: The layers in `areas` are summarized within the entire catchment for each reach
}
