% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{WSCheckDrainage}
\alias{WSCheckDrainage}
\title{Check and fix problems with drainage direction}
\usage{
WSCheckDrainage(connMat, drainMat, prevProbs = NA)
}
\arguments{
\item{connMat}{preliminary connectivity matrix}

\item{drainMat}{Drainage direction matrix}

\item{prevProbs}{Previous number of problems, to allow stopping if no improvement on 
subsequent calls}
}
\value{
A corrected connectivity matrix
}
\description{
Check and fix problems with drainage direction
}
\details{
In some cases, drainage direction rasters don't agree with flow accumulation, resulting
		in a delineated stream that doesn't have the right drainage direction. This function
		attempts to detect and fix this in the adjacency matrix and the drainage layer
}
\keyword{internal}
