% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{WSComputeLength}
\alias{WSComputeLength}
\title{Compute length to next pixel given drainage direction}
\usage{
WSComputeLength(drainage, cellsize)
}
\arguments{
\item{drainage}{drainage direction vector}

\item{cellsize}{size of each cell (vector of length 2)}
}
\value{
vector of lengths
}
\description{
Compute length to next pixel given drainage direction
}
\keyword{internal}
