% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{WSConnectivity}
\alias{WSConnectivity}
\title{Compute connectivity matrix}
\usage{
WSConnectivity(drainage, stream)
}
\arguments{
\item{drainage}{Drainage direction raster}

\item{stream}{Stream network raster; see `details`}
}
\value{
A [Matrix::sparseMatrix()] representation of the river network. For a `stream` input
	raster with `n` non-NA cells, the dimensions of this matrix will be n by n. Row/column
	names of the matrix will be the pixel IDs from the `stream` input raster. Values of the
	matrix cells are either 0 or 1; a zero indicates no flow, a one in cell i,j indicates
	that reach `i` receives water from reach `j`.
}
\description{
Compute connectivity matrix
}
\details{
The stream network raster should be NA in all cells not considered a part of the
	river network. The pixel values of the raster must be unique IDs representing individual
	stream reaches to model. At present, the only supported reach size is a single pixel, thus
	each pixel must have a unique value.
}
\keyword{internal}
