% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{wsDistance}
\alias{wsDistance}
\title{Produces a site by pixel distance matrix}
\usage{
wsDistance(ws, x, variable = "length")
}
\arguments{
\item{ws}{A watershed}

\item{x}{A vector of sites}

\item{variable}{The variable to use as the distance}
}
\value{
A site by pixel distance matrix
}
\description{
Note that pixels in the watershed that are not connected to any of the sites will be excluded
from the result.
}
