% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{WSFixDrainage}
\alias{WSFixDrainage}
\title{Fix drainage direction for a single pixel}
\usage{
WSFixDrainage(id, drainMat, connMat)
}
\arguments{
\item{id}{ID of the problematic pixel}

\item{drainMat}{Drainage direction matrix}

\item{connMat}{preliminary connectivity matrix}
}
\description{
Fix drainage direction for a single pixel
}
\keyword{internal}
