% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{WSFlowTo}
\alias{WSFlowTo}
\title{Compute which pixels flow into which other pixels}
\usage{
WSFlowTo(mat)
}
\arguments{
\item{mat}{A matrix with minimum three columns, the first being the x-coordinate, second the y,
and third the ID.}
}
\value{
A matrix of IDs, the first column the source, the second column the destination
}
\description{
Compute which pixels flow into which other pixels
}
\keyword{internal}
