% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_water_yr.R
\name{wtr_yr}
\alias{wtr_yr}
\title{Add Water Year}
\usage{
wtr_yr(dates, start_month = 10)
}
\arguments{
\item{dates}{at Date formatted column in a dataframe}

\item{start_month}{defaults to Oct 1, the starting month}
}
\description{
This function takes a \emph{date} column and
returns the Water Year (from Oct 1 to Sep 30).
Can specify the starting month for a given water year.
}
\details{
Details: Here
}
