% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{accumulate}
\alias{accumulate}
\title{Returns all points in the watershed connecting two points, accumulating a value between}
\usage{
accumulate(ws, upstream, downstream = Inf, direction = c("down", "up"),
  variable = "length", parallel = FALSE)
}
\arguments{
\item{ws}{Watershed object}

\item{upstream}{ID number of upstream point; if Inf ALL upstream pixels will be returned}

\item{downstream}{ID of downstream point; if Inf ALL downstream pixels will be returned}

\item{direction}{In which direction should the variable be accumulated; see 'details'}

\item{variable}{The variable to accumulate}

\item{parallel}{Boolean, should parallel processing be used?}
}
\value{
A matrix, first column connected pixelIDs, second the accumulated variable
}
\description{
If direction is down, then the accumulated distances will reflect the distance from the 
`upstream` point to each downstream point, and values will be positive. If direction is "up",
then the distances will be negative and accumulate from `downstream` to `upstream`.
}
