% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_WYD.R
\name{add_WYD}
\alias{add_WYD}
\title{Add Day of Year, Day of Water Year.}
\usage{
add_WYD(df, datecolumn)
}
\arguments{
\item{datecolumn}{requires a POSIXct formatted date or datetime}

\item{dates}{a dataframe}
}
\description{
This function takes a POSIXct formatted column name and
adds new columns to a dataframe for Day of Year (Julian),
Day of Water Year (starting Oct 1 and ending Sep 30),
and Water Year (from Oct 1 to Sep 30).
}
\details{
Here's the specific details:
}
