% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_wytypes}
\alias{ca_wytypes}
\title{California Water Year Types.}
\format{
A data frame with seven variables: \code{WY}, \code{Oct_Mar},
  \code{Apr_Jul}, \code{WYsum}, \code{Index}, \code{WYtype}, and
  \code{watershed}. These are the official Water Types for CA.
}
\usage{
ca_wytypes
}
\description{
Water Year Types from the CDEC website for San Joaquin and Sacramento Indices. See \url{http://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}
}
\keyword{datasets}
