% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{connect}
\alias{connect}
\title{Returns all points in the watershed connecting two points}
\usage{
connect(ws, upstream, downstream = Inf)
}
\arguments{
\item{ws}{Watershed object}

\item{upstream}{ID number of upstream point; if Inf ALL upstream pixels will be returned}

\item{downstream}{ID of downstream point; if Inf ALL downstream pixels will be returned}
}
\value{
A vector of pixel ids
}
\description{
This is a simple wrapper to accumulate that only returns the pixels
}
