% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{connect}
\alias{connect}
\title{Returns all points in the watershed connecting two points}
\usage{
connect(ws, point1, point2)
}
\arguments{
\item{ws}{Watershed object}

\item{point1}{ID number of upstream point}

\item{point2}{ID of downstream point}
}
\value{
vector of pixel ids
}
\description{
Returns all points in the watershed connecting two points
}
