% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connectCPP}
\alias{connectCPP}
\title{Connect points in a watershed}
\usage{
connectCPP(dsPixel, upstream, downstream)
}
\arguments{
\item{dsPixel}{A vector of downstream pixels, diPixel[i] is downstream from pixel i}
}
\description{
Connect points in a watershed
}
