% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{discharge_scaling}
\alias{discharge_scaling}
\title{Compute discharge from catchment area}
\usage{
discharge_scaling(A, calib)
}
\arguments{
\item{A}{vector of catchment area, in \eqn{m^2}}

\item{calib}{A data.frame with two elements, `A` (catchment area) and `Q` (observed discarge)}
}
\description{
Compute discharge from catchment area
}
\details{
Computes discharge from catchment area as \eqn{\log Q = \log b + m \log A}.
		If a single point in `calib` is included, the relationship will be re-parameterised by 
 	adjusting the intercept parameter `b` so that the calibration point falls on the line  
		(while keeping the slope the same).

		With multiple points, a bayesian regression model is fit with rstanarm. The model uses
		Burgers et al parameters as an informative prior.

		The default parameters used by this function come from Burgers et al (2014). 
		For these parameters, catchment area units are expected to be in 
		\eqn{m^2}, and discharge will be computed in \eqn{m^3 s^{-1}}.
}
\references{
Burgers HE et al. 2014. Size relationships of water discharge in rivers: scaling
		of discharge with catchment area, main-stem lengthand precipitation. 
		*Hydrological Processes*. **28**:5769-5775.
}
