% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{outlets}
\alias{outlets}
\title{Get data from all outlets of a watershed}
\usage{
outlets(ws, rid = NA, output = c("data.frame", "Spatial"))
}
\arguments{
\item{ws}{Watershed object}

\item{rid}{vector of reach IDs, if NA returns outlet for entire network}

\item{output}{Output type to return}
}
\value{
a `data.frame` or a `SpatialPixelsDataFrame` containing data for all outlets
}
\description{
Get data from all outlets of a watershed
}
