% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{rasterToSPDF}
\alias{rasterToSPDF}
\title{Produce a [sp::SpatialPixelsDataFrame] from a [raster::RasterLayer]}
\usage{
rasterToSPDF(x)
}
\arguments{
\item{x}{A [raster::RasterLayer] object}
}
\value{
A [sp::SpatialPixelsDataFrame]
}
\description{
Produce a [sp::SpatialPixelsDataFrame] from a [raster::RasterLayer]
}
\keyword{internal}
