% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{splitReaches}
\alias{splitReaches}
\title{Splits a reachID at selected points}
\usage{
splitReaches(ws, points, na_ignore = FALSE)
}
\arguments{
\item{ws}{A watershed object}

\item{points}{A vector of ids at which to split reaches}

\item{na_ignore}{Logical, if `TRUE` NAs will be dropped from the data. If `FALSE`, any NAs
in `points` will cause an error.}
}
\value{
A modified watershed with new reachIDs
}
\description{
Splits a reachID at selected points
}
