% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\name{transport}
\alias{transport}
\title{Run the transport model}
\usage{
transport(adjacency, initial, steps, weights)
}
\arguments{
\item{adjacency}{The adjacency matrix describing the stream network}

\item{initial}{A vector of initial conditions}

\item{steps}{How may time steps to run the model}

\item{weights}{Optional weighting to apply to the adjacency matrix}
}
\description{
Run the transport model
}
