% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{ws_intersect}
\alias{ws_intersect}
\title{Intersect a river feature with a polygon feature}
\usage{
ws_intersect(ws, ...)
}
\arguments{
\item{ws}{A Watershed}

\item{...}{Additional parameters for [watershed::w_intersect()]}
}
\value{
A data.table summarising the polygons in each layer in areas along the river provided in x.
}
\description{
`Watershed`-friendly wrapper to [watershed::w_intersect()], see the help file in the 
`watershed` package for details on the function.
}
