% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_intersect.R
\name{wsi_summary}
\alias{wsi_summary}
\alias{.wsi_summary}
\alias{.wsi_summary_single}
\title{Summarise multiple raster layers by area}
\usage{
.wsi_summary(ras, ...)

.wsi_summary_single(ras)
}
\arguments{
\item{ras}{A raster stack}

\item{...}{Additional (single-valued) attributes to add to the output table}
}
\value{
A data.table with the following columns:
 * method: The method used, either 'riparian', 'riparian_upstream', or 'catchment'
 * layer: The layer of origin for each row
 * category: The numeric category from the raster layer of each row
 * area: The area occupied by the category
 * proportion: The proportion of area (within the method and layer)
}
\description{
Summarise multiple raster layers by area
}
\keyword{internal}
